requirements.txt
openai>=1.51.0
rouge-score>=0.1.2
sacrebleu>=2.4.0
bert-score>=0.3.13
python-dotenv>=1.0.1